Add simple test for OCI on multi-node/multi-gpu runtime
The new test ensures that OCI support in multi-node/multi-gpu is working properly. This is done by deploying basic InferenceService with an OCI image in modelcar format.

Since the storage is the only different variable, at the moment, we would trust the tests in test_nvidia_multi_node.py are providing proper coverage for the rest of the features.